requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
requests==2.28.1
smmap==5.0.0
urllib3==1.26.12
